Restructure component and storybook files
