+++ src/components/Label/Label.stories.tsx
