+++ src/stories/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Header } from './Header';
+// import { Header } from './Header';
 import './page.css';
 
 type User = {
@@ -12,12 +12,12 @@ export const Page: React.VFC = () => {
 
   return (
     <article>
-      <Header
+      {/* <Header
         user={user}
         onLogin={() => setUser({ name: 'Jane Doe' })}
         onLogout={() => setUser(undefined)}
         onCreateAccount={() => setUser({ name: 'Jane Doe' })}
-      />
+      /> */}
 
       <section>
         <h2>Pages in Storybook</h2>
@@ -36,7 +36,7 @@ export const Page: React.VFC = () => {
         <ul>
           <li>
             Use a higher-level connected component. Storybook helps you compose such data from the
-            "args" of child component stories
+            args of child component stories
           </li>
           <li>
             Assemble data in the page component from your services. You can mock these services out
